+++ PrePost/Reference/realInferenceParameterStmt.free.pre.post.f90
@@ -18,7 +18,7 @@ integer :: dp
 parameter (dp = kind(1.0d0))
 real(dp) :: y
 integer :: i,j
-  real(kind=dp) :: oad_ctmp0
+  double precision :: oad_ctmp0
 i=2
 j=3
 oad_ctmp0 = real(i+1,kind=dp)/real(j-1,kind=dp)
